Added Linux 64 bit fixes to the game code.
git-svn-id: svn://0x263a.de/tekuum_src@213 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ game/ai/AAS_routing.cpp
@@ -167,7 +167,10 @@ void idAASLocal::CalculateAreaTravelTimes(void) {
 		}
 	}
 
-	assert( ( (unsigned int) bytePtr - (unsigned int) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
+	// Techyon RB: 64 bit fixes, changed unsigned int to ptrdiff_t
+	assert( ( (ptrdiff_t) bytePtr - (ptrdiff_t) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
+	// Techyon END
+
 }
 
 /*
